+++ src/test/java/com/svetylkovo/rojo/matcher/RojoBeanProcessorTest.java
@@ -7,6 +7,11 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
 
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
+import java.util.Date;
+import java.util.List;
+
 import static org.junit.Assert.*;
 
 public class RojoBeanProcessorTest {
@@ -40,58 +45,25 @@ public class RojoBeanProcessorTest {
 
     @Test
     public void noFlagsTest() {
-        RojoBeanProcessor<NoFlagsBean> processor = new RojoBeanProcessor<>(NoFlagsBean.class);
-        processor.processAnnotations();
-
-        MatchIterator it = new MatchIterator(processor.getMatcher(flagsTestString));
-
-        BeanIterator<NoFlagsBean> beanIterator = new BeanIterator<>(it, processor);
-        beanIterator.hasNext();
-
-        NoFlagsBean bean = beanIterator.next();
+        NoFlagsBean bean = getMatchedBean(NoFlagsBean.class, flagsTestString);
         assertEquals("abc,def,", bean.getValue());
     }
 
     @Test
     public void flagsTest() {
-        RojoBeanProcessor<FlagsBean> processor = new RojoBeanProcessor<>(FlagsBean.class);
-        processor.processAnnotations();
-
-        MatchIterator it = new MatchIterator(processor.getMatcher(flagsTestString));
-
-        BeanIterator<FlagsBean> beanIterator = new BeanIterator<>(it, processor);
-        beanIterator.hasNext();
-
-        FlagsBean bean = beanIterator.next();
+        FlagsBean bean = getMatchedBean(FlagsBean.class, flagsTestString);
         assertEquals("abc,def,\nghi,jkl", bean.getValue());
     }
 
     @Test
     public void mapperTest() {
-        RojoBeanProcessor<MapperBean> processor = new RojoBeanProcessor<>(MapperBean.class);
-        processor.processAnnotations();
-
-        MatchIterator it = new MatchIterator(processor.getMatcher("123"));
-
-        BeanIterator<MapperBean> beanIterator = new BeanIterator<>(it, processor);
-        beanIterator.hasNext();
-
-        MapperBean bean = beanIterator.next();
-
+        MapperBean bean = getMatchedBean(MapperBean.class, "123");
         assertArrayEquals(new int[]{1,2,3}, bean.getNumbers());
     }
 
     @Test
     public void nestedTest() {
-        RojoBeanProcessor<NestedMain> processor = new RojoBeanProcessor<>(NestedMain.class);
-        processor.processAnnotations();
-
-        MatchIterator it = new MatchIterator(processor.getMatcher("a:123"));
-
-        BeanIterator<NestedMain> beanIterator = new BeanIterator<>(it, processor);
-        beanIterator.hasNext();
-
-        NestedMain bean = beanIterator.next();
+        NestedMain bean = getMatchedBean(NestedMain.class, "a:123");
         NestedInner inner = bean.getInner();
 
         assertEquals("a", bean.getLetter());
@@ -102,19 +74,69 @@ public class RojoBeanProcessorTest {
 
     @Test
     public void nestedWrongMatchTest() {
-        RojoBeanProcessor<NestedMain> processor = new RojoBeanProcessor<>(NestedMain.class);
-        processor.processAnnotations();
-
-        MatchIterator it = new MatchIterator(processor.getMatcher("a:xxx"));
-
-        BeanIterator<NestedMain> beanIterator = new BeanIterator<>(it, processor);
-        beanIterator.hasNext();
-
-        NestedMain bean = beanIterator.next();
+        NestedMain bean = getMatchedBean(NestedMain.class, "a:xxx");
         NestedInner inner = bean.getInner();
 
         assertEquals("a", bean.getLetter());
         assertNull(inner);
+    }
+
+    @Test
+    public void listOfFieldsPlainTest() {
+        ListBeanPlain bean = getMatchedBean(ListBeanPlain.class, "a:123");
+        List<Integer> numbers = bean.getNumbers();
+
+        assertEquals("a", bean.getLetter());
+        assertEquals(1, numbers.get(0).intValue());
+        assertEquals(2, numbers.get(1).intValue());
+        assertEquals(3, numbers.get(2).intValue());
+    }
+
+    @Test
+    public void listOfFieldsBadTest() {
+        expectedEx.expect(RuntimeException.class);
+        expectedEx.expectMessage("Could't get the regexp pattern for the List<> field numbers in the "+ListBeanBad.class.getName()+" class. Use either @Regex to annotate the List<> field itself, or use the List whose generic type is of a class, which is annotated by the @Regex and @Group accordingly.");
+        ListBeanBad bean = getMatchedBean(ListBeanBad.class, "a:123");
+    }
+
+    @Test
+    public void listOfFieldsNestedTest() {
+        ListBeanNested bean = getMatchedBean(ListBeanNested.class, "a:123456789");
+        List<NestedInner> numbers = bean.getNumbers();
+
+        assertEquals("a", bean.getLetter());
+
+        int expected = 1;
+        for ( NestedInner num : numbers) {
+            assertEquals(expected++, num.getFirst());
+            assertEquals(expected++, num.getSecond());
+            assertEquals(expected++, num.getThird());
+        }
+    }
+
+    @Test
+    public void listOfFieldsDateTest() throws ParseException {
+        ListBeanDate bean = getMatchedBean(ListBeanDate.class, "a:2000/2001/2002");
+        List<Date> years = bean.getYears();
+
+        assertEquals("a", bean.getLetter());
+
+        SimpleDateFormat sdf = new SimpleDateFormat("yyyy");
+        assertEquals(sdf.parse("2000"), years.get(0));
+        assertEquals(sdf.parse("2001"), years.get(1));
+        assertEquals(sdf.parse("2002"), years.get(2));
+    }
+
+    private <T> T getMatchedBean(Class<T> clazz, String input) {
+        RojoBeanProcessor<T> processor = new RojoBeanProcessor<>(clazz);
+        processor.processAnnotations();
+
+        MatchIterator it = new MatchIterator(processor.getMatcher(input));
+
+        BeanIterator<T> beanIterator = new BeanIterator<>(it, processor);
+        beanIterator.hasNext();
+
+        return beanIterator.next();
     }
 
 }